Remove back-links from the journey

+++ src/handlers/save.js
@@ -12,13 +12,17 @@ module.exports = class SaveHandler extends BaseHandler {
   }
 
   /**
-   * Get handler for select year page
+   * If external user zap the session and display the save page. If a
+   * fish management team user then just go back to teh licence selector
    * @param request
    * @param h
    * @param user
    * @returns {Promise<*>}
    */
   async doGet (request, h) {
+    if (process.env.CONTEXT === 'FMT') {
+      return h.redirect('/licence')
+    }
     await request.cache().drop()
     request.cookieAuth.clear()
     return h.view(this.path)
